I already created this, so i just uploaded it to git.
This website is one of the assignment we build while in school and i just decided to keep it in my git account for future purpose. The main thing is that it was written with Bootstrap, PHP, Mysql.
I did a contact page and a blog page(a working one).